+++ changelog.md
@@ -51,6 +51,9 @@
 - react-native-bpk-component-banner-alert: 1.1.13 => 1.2.0
   - New animations available when `show` is toggled. See https://backpack.github.io/components/native/banner-alerts
 
+- bpk-component-datatable:
+  - New data table component, see https://backpack.github.io/components/web/datatables
+
 ## 2017-12-06 - More theming support
 
 **Added:**